Port Seagull's command line handling
This commit adds a loose port of the command line parsing found in Seagull with
slight modification to make it work under Python 2.7. The basic commands found
in Seagull are ported over, including:

- configuration options
- process options (background, etc, although not implmented in the App class
  yet)
- assets management with full support for skins
- assets commands for starting/stopping watchers and recompiling
- commands for creating custom configuration and skin directories

As a result of these changes, the skin directory is now in
broadcast/skins/broadcast, and a few options are added to the default
configuration file for skin set-up. Bundles are now switched to YAML-like
format used in Seagull, and are no longer listed in the ini. The --quiet switch
is added which silences the console logging.

A new system for managing hooks was added which collects all hooks during
initialization, and runs them in a separate call where appropriate. This allows
different hooks to add new hooks via the exposed add_hook() method on the exts
container.

A new module is added in util.skinning whihc provides functions for getting
skin directories.

Signed-off-by: Branko Vukelic <branko@outernet.is>