Implement accept invitation
This commit adds the AcceptInvitation route which is used to accept beta
invitation. It is coded in a way that allows later usage as generic invitation
mechanism. To generalize token handling, ConfirmationMixin class was added, and
some of the routes were moved to this mixin (e.g., email confirmation, password
reset).

The auth-related forms and models were also fixed to support the implemented
features.

Signed-off-by: Branko Vukelic <branko@outernet.is>