Set up iscandidate func for every query
Because the iscandidate func is not idempotent, we need to set it up over and
over again for every connection. This patch modfies the ContentItem model so
that it creates a new connection for every method that uses the iscandidate
function, and adds the function to the new connection. This is expected to have
a slight performance penalty but ensures correctness of the query.

Signed-off-by: Branko Vukelic <branko@outernet.is>