Only use the loader the first time
When reloading an existing ROCA-loaded container, the page layout breaks in
unexpected ways and creates a somewhat unpleasant scroll-all-over effect. This
patch modifies the roca plugin such that it only uses the loader the first time
it loads the content. On reloads, it will not replace the existing contents
until the new content is fetched.

Signed-off-by: Branko Vukelic <branko@outernet.is>