Add route mixin and command for working w/ API ACL
This commit adds a new command, 'apiuser', and APIMixin route mixin. Both rely
on the newly introduced app.api_acl option, which points to a (not necessarily
existing) JSON file that would contain the ACL for API calls.

The API endpoints are expected to use APIMixin as part of the class definition.
The authentication is, therefore, completely separate from the normal user
database.

Credentials for API calls is created using the 'apiuser' command. The command
takes only username as its argument, and will generate a random 32-digit hex
number as the access token / password.

Signed-off-by: Branko Vukelic <branko@outernet.is>