Implement bin finalizing and bin-related API
This commit modifies the ContentItem and Bin models to support bin finalization
and features required for retrieval of bin data. 4 API endpoints are also added
via the broadcast.routes.api module, which allow finalizing the bin, retrieval
of bin list, retrieval of item lists for bins (by ID or the last one), and
retrieval of file data for individual items.

Signed-off-by: Branko Vukelic <branko@outernet.is>