Add ability to force state update
This commit modifies the last-update module to add the ability to force state
update and suspend the schedule state update while forced update is taking
place. This reduces the possibility of a rance condition where post-vote update
to the page content overlaps with state-update-indicuded reloads. Most notably,
this solves the issue where post-vote highlight disappears as the highlighted
item is cleared by state-update-induced reload.

Whether the state update is forced or not is indicated by the new 'forced'
attribute on the data attached to the state-update event. This is used in the
main module to suppress the feedback pop-up.

The voting module has been updated to take advantage of the forced updates.

Signed-off-by: Branko Vukelic <branko@outernet.is>