Use CBRH for errors, convert sess/auth to hooks
This commit converts the error handler to CBRH, and sessions and auth plugins
to streamline before and after hooks. This is done because the bottle error
handlers are not using the hooks needed to make the user object available to
the template which are used in the error template, and are also not subjet to
plugins.

Fixes #43

Signed-off-by: Branko Vukelic <branko@outernet.is>